[CHORE] homeController 분리
