+++ src/controller/index.ts
@@ -1,1 +1,2 @@
 export { default as goalController } from "./goalController";
+export { default as homeController } from "./homeController";
